Add functional basic math operation on display
This commit is responsible for adding event listeners on operators and equal sign. Buttons assign it's operator to operator variable. Equal button runs operate function and resets number2 value to empty string after that. Also there is a change to basic mathematical functions. Now they calculate a result which is shown on the display, then assign the result to variable number1. There are also changes to operators classes in index.html, so they can be chosen separately from the rest of buttons.