Add keyboard support
This commit is responsible for adding working keyboard support. It achieves that thanks to 'keydown' event listener. All previous working 'click' event listeners are transformed into additional 'keydown' listeners. Also this commit changes DELETE button to BACKSPACE, which works the same. This is a readability change. Now button name is the same as keyboard button.